refactor: exposed EntityActionType to public API
