+++ src/lib/actions/error.ts
@@ -1,6 +1,7 @@
-import { EntityActionType, generateActionObject } from '../internal';
+import { generateActionObject } from '../internal';
 import { EntityState } from '../entity-state';
 import { Type } from '@angular/core';
+import { EntityActionType } from '@ngxs-labs/entity-state';
 
 export interface EntitySetErrorAction {
   payload: Error;
